genesis_block_explorer/models/genesis/automap.py
from sqlalchemy import create_engine, MetaData, ForeignKey, Column, Integer
from sqlalchemy.orm import sessionmaker, Session
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.ext.automap import automap_base

from dictalchemy import make_class_dictable

from ...db import db

from ..db_engine.automap import automapped_classes
