requirements.txt
wheel
sqlalchemy-datatables
Flask
jinja2
Flask-SQLAlchemy
Flask-RESTful
Flask-JWT
psycopg2-binary
inflect
dictalchemy
bumpversion
six
pytz
aniso8601
jwt
Flask-JsonTools
msgpack
attrdict
